Re VIV-322, VIV-343: Return schema default instead of None for unknown ICommonMetadata fields of free teasers
Clients rely on the fact that iterable fields always are iterable, or
rather that ICommonMetadata behaves like the CommonMetadata base
class, and I think that's the right idea.

Bluntly returning None breaks that assumption, and creates confusion,
since the authorships property of CommonMetadata always returns an
empty tuple, never None.